+++ fe/src/pages/index.js
@@ -1,6 +1,6 @@
 import React, { Fragment } from "react";
 import Guest from "./Authentication/Guest";
-import UserDashboard from "./user/Home/index";
+import UserDashboard from "./user/Home/Home";
 import AdminDashboard from "./admin/index";
 import { getUserToken, isUser } from "../utils";
